Changgelog: Added development tests (test_sprint4.py) to test some of the functionality of this final sprint and assure all methods are in working order.

+++ tests/test_sprint3.py
@@ -17,7 +17,7 @@ def new_table(cursor: sqlite3.Cursor):
     cursor.execute('''CREATE TABLE IF NOT EXISTS dummy_pop(
                    id TEXT PRIMARY KEY,
                    rank TEXT,
-                   rankUpDown FLOAT,
+                   rankUpDown INTEGER,
                    title TEXT,
                    fullTitle INTEGER,
                    year FLOAT DEFAULT 0,
